Split up form into sections
