+++ ui/src/pages/deployment-details/deployment-details-form/config.ts
